refactor(ai_visibility): struct keys + output-derived labels (#406)
AiSignal/DimensionScore + chunk sections/recommendation + KG suggestions carry
canonical kind enums (+ values); all text baked English (JSON canonical).
Per-area pure text fns are the single source. render_ai_visibility re-derives
in the run locale for the PDF. Also fixes two redundant-cast clippy warnings.

+++ src/content_visibility/mod.rs
@@ -852,7 +852,7 @@ fn build_organic_visibility(seo: &crate::seo::SeoAnalysis) -> Vec<ContentSignal>
                 AssessmentLevel::Warning,
                 EvidenceConfidence::High,
                 H1Multiple,
-                ContentSignalValues::count(n as usize),
+                ContentSignalValues::count(n),
             )
             .with_evidence(
                 ContentEvidence::new(EvidenceSource::AxTree, EvidenceConfidence::High)
